Extract combo and accent layer settings
