+++ Assets/_Project/Scripts/RazerChroma/AccentLayerSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 54173228310d4328ac6203abfaaf82c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
